+++ packages/core/src/client.ts
@@ -1,5 +1,7 @@
 import {
   checkResourcesResponseFromProtobuf,
+  deleteSchemasResponseFromProtobuf,
+  disablePoliciesResponseFromProtobuf,
   getPoliciesResponseFromProtobuf,
   getSchemasResponseFromProtobuf,
   listPoliciesResponseFromProtobuf,
@@ -11,6 +13,7 @@ import {
   addOrUpdateSchemasRequestToProtobuf,
   checkResourcesRequestToProtobuf,
   deleteSchemasRequestToProtobuf,
+  disablePoliciesRequestToProtobuf,
   getPoliciesRequestToProtobuf,
   getSchemasRequestToProtobuf,
   planResourcesRequestToProtobuf,
@@ -25,6 +28,9 @@ import type {
   CheckResourcesResponse,
   CheckResourcesResult,
   DeleteSchemasRequest,
+  DeleteSchemasResponse,
+  DisablePoliciesRequest,
+  DisablePoliciesResponse,
   GetPoliciesRequest,
   GetPoliciesResponse,
   GetSchemasRequest,
@@ -288,17 +294,22 @@ export abstract class Client {
    *
    * - the client to be configured with {@link Options.adminCredentials},
    *
-   * - the Cerbos policy decision point server to be configured with the {@link https://docs.cerbos.dev/cerbos/latest/api/admin_api.html | admin API} enabled, and
+   * - the Cerbos policy decision point (PDP) server to be configured with the {@link https://docs.cerbos.dev/cerbos/latest/api/admin_api.html | admin API} enabled, and
    *
    * - a dynamic {@link https://docs.cerbos.dev/cerbos/latest/configuration/storage.html | storage backend}.
    *
+   * The way this method handles failure depends on the version of the connected PDP server.
+   * When the server is running Cerbos v0.25 or later, it returns `true` if the schema was deleted and `false` if the schema was not found.
+   * With earlier versions of Cerbos, it throws an error if the schema was not found, and returns successfully if the schema was deleted; the returned value should be ignored.
+   *
    * @example
    * ```typescript
-   * await cerbos.deleteSchema("document.json");
+   * const deleted = await cerbos.deleteSchema("document.json");
    * ```
    */
-  public async deleteSchema(id: string): Promise<void> {
-    await this.deleteSchemas({ ids: [id] });
+  public async deleteSchema(id: string): Promise<boolean> {
+    const { deletedSchemas } = await this.deleteSchemas({ ids: [id] });
+    return deletedSchemas === 1;
   }
 
   /**
@@ -309,19 +320,79 @@ export abstract class Client {
    *
    * - the client to be configured with {@link Options.adminCredentials},
    *
-   * - the Cerbos policy decision point server to be configured with the {@link https://docs.cerbos.dev/cerbos/latest/api/admin_api.html | admin API} enabled, and
+   * - the Cerbos policy decision point (PDP) server to be configured with the {@link https://docs.cerbos.dev/cerbos/latest/api/admin_api.html | admin API} enabled, and
+   *
+   * - a dynamic {@link https://docs.cerbos.dev/cerbos/latest/configuration/storage.html | storage backend}.
+   *
+   * The way this method handles failure depends on the version of the connected PDP server.
+   * When the server is running Cerbos v0.25 or later, it returns a {@link DeleteSchemasResponse} that includes the number of schemas that were deleted.
+   * With earlier versions of Cerbos, it throws an error if no schemas were found, and returns successfully if at least one schema was deleted; the returned value should be ignored.
+   *
+   * @example
+   * ```typescript
+   * const result = await cerbos.deleteSchemas({
+   *   ids: ["document.json", "image.json"],
+   * });
+   * ```
+   */
+  public async deleteSchemas(
+    request: DeleteSchemasRequest
+  ): Promise<DeleteSchemasResponse> {
+    return deleteSchemasResponseFromProtobuf(
+      await this.admin("deleteSchema", deleteSchemasRequestToProtobuf(request))
+    );
+  }
+
+  /**
+   * Disable multiple policies.
+   *
+   * @remarks
+   * Requires
+   *
+   * - the client to be configured with {@link Options.adminCredentials},
+   *
+   * - the Cerbos policy decision point server to be at least v0.25 and configured with the {@link https://docs.cerbos.dev/cerbos/latest/api/admin_api.html | admin API} enabled, and
    *
    * - a dynamic {@link https://docs.cerbos.dev/cerbos/latest/configuration/storage.html | storage backend}.
    *
    * @example
    * ```typescript
-   * await cerbos.deleteSchemas({
-   *   ids: ["document.json", "image.json"],
+   * const result = await cerbos.disablePolicies({
+   *   ids: ["resource.document.v1", "resource.image.v1"],
    * });
    * ```
    */
-  public async deleteSchemas(request: DeleteSchemasRequest): Promise<void> {
-    await this.admin("deleteSchema", deleteSchemasRequestToProtobuf(request));
+  public async disablePolicies(
+    request: DisablePoliciesRequest
+  ): Promise<DisablePoliciesResponse> {
+    return disablePoliciesResponseFromProtobuf(
+      await this.admin(
+        "disablePolicy",
+        disablePoliciesRequestToProtobuf(request)
+      )
+    );
+  }
+
+  /**
+   * Disable a policy.
+   *
+   * @remarks
+   * Requires
+   *
+   * - the client to be configured with {@link Options.adminCredentials},
+   *
+   * - the Cerbos policy decision point server to be at least v0.25 and configured with the {@link https://docs.cerbos.dev/cerbos/latest/api/admin_api.html | admin API} enabled, and
+   *
+   * - a dynamic {@link https://docs.cerbos.dev/cerbos/latest/configuration/storage.html | storage backend}.
+   *
+   * @example
+   * ```typescript
+   * const disabled = await cerbos.disablePolicy("resource.document.v1");
+   * ```
+   */
+  public async disablePolicy(id: string): Promise<boolean> {
+    const { disabledPolicies } = await this.disablePolicies({ ids: [id] });
+    return disabledPolicies === 1;
   }
 
   /**