+++ packages/core/src/convert/fromProtobuf.ts
@@ -19,6 +19,8 @@ import type {
 import type {
   CheckResourcesResponse as CheckResourcesResponseProtobuf,
   CheckResourcesResponse_ResultEntry,
+  DeleteSchemaResponse,
+  DisablePolicyResponse,
   GetPolicyResponse,
   GetSchemaResponse,
   ListPoliciesResponse as ListPoliciesResponseProtobuf,
@@ -35,6 +37,7 @@ import type {
   Condition,
   DerivedRoleDefinition,
   DerivedRoles,
+  DisablePoliciesResponse,
   GetPoliciesResponse,
   ListPoliciesResponse,
   ListSchemasResponse,
@@ -67,6 +70,7 @@ import {
   SchemaDefinition,
   ValidationErrorSource,
 } from "../types/external";
+import type { DeleteSchemasResponse } from "../types/external/DeleteSchemasResponse";
 import type { GetSchemasResponse } from "../types/external/GetSchemasResponse";
 import type { OmitFromEach } from "../types/internal";
 
@@ -138,6 +142,18 @@ const validationErrorSourceFromProtobuf = (
       );
   }
 };
+
+export const deleteSchemasResponseFromProtobuf = ({
+  deletedSchemas,
+}: DeleteSchemaResponse): DeleteSchemasResponse => ({
+  deletedSchemas,
+});
+
+export const disablePoliciesResponseFromProtobuf = ({
+  disabledPolicies,
+}: DisablePolicyResponse): DisablePoliciesResponse => ({
+  disabledPolicies,
+});
 
 export const getPoliciesResponseFromProtobuf = ({
   policies,